python-scripts/requirements.txt
discord
steam
dota2
python-dotenv
d2api
PyYAML
more-itertools